fix: Logout does not fail when user is already logged out

+++ internal/users/service/user_service.go
@@ -140,5 +140,9 @@ func (s *userServiceImpl) Logout(ctx context.Context, id uuid.UUID) error {
 		return err
 	}
 
+	if len(apiKeys) == 0 {
+		return nil
+	}
+
 	return s.apiKeyRepo.Delete(ctx, apiKeys)
 }
